Feat: 동영상 제외한 로직 연결
